remove SCP-liveness-prepare-3.ivy
Given the ordering axioms for ballots it is not the case that a set of
ballots always has a maximum. Thus SCP-liveness-prepare-3 is pointless.
What we could do is have nodes track the max confirmed prepared ballot
they saw in a local variable and use that in the change_ballot action.